requirements.txt
distlib==0.3.7
filelock==3.12.4
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.21.12
pydub==0.25.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
pytube==15.0.0
reportlab==4.0.7
six==1.16.0
virtualenv==20.24.5
